backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn==0.25.0

# Database
pymongo==4.5.0
motor==3.3.1

# Validation & Types
pydantic>=2.6.4
email-validator>=2.2.0

# Environment
python-dotenv>=1.0.1

# File uploads
python-multipart>=0.0.9

# HTTP requests
requests>=2.31.0

# Timezone data
tzdata>=2024.2
